Rework the toolbar on the loadout page to use the styling on the downloads page
