+++ src/NexusMods.App.UI/NexusMods.App.UI.csproj
@@ -158,12 +158,6 @@
     <Compile Update="RightContent\LoadoutGrid\Columns\ModCategoryViewModel.cs">
       <DependentUpon>IModCategoryViewModel.cs</DependentUpon>
     </Compile>
-    <Compile Update="Toolbars\DefaultLoadoutToolbarDesignViewModel.cs">
-      <DependentUpon>IDefaultLoadoutToolbarViewModel.cs</DependentUpon>
-    </Compile>
-    <Compile Update="Toolbars\DefaultLoadoutToolbarViewModel.cs">
-      <DependentUpon>IDefaultLoadoutToolbarViewModel.cs</DependentUpon>
-    </Compile>
     <Compile Update="RightContent\Downloads\CompletedDesignViewModel.cs">
       <DependentUpon>ICompletedViewModel.cs</DependentUpon>
     </Compile>
