+++ src/NexusMods.App.UI/RightContent/LoadoutGrid/ILoadoutGridViewModel.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.ObjectModel;
 using NexusMods.App.UI.RightContent.LoadoutGrid.Columns;
-using NexusMods.App.UI.Toolbars;
 using NexusMods.DataModel.Loadouts;
 using NexusMods.DataModel.Loadouts.Cursors;
 
@@ -11,11 +10,12 @@ namespace NexusMods.App.UI.RightContent.LoadoutGrid;
 /// </summary>
 public interface ILoadoutGridViewModel : IRightContentViewModel
 {
-    public ILoadoutToolbarViewModel Toolbar { get; }
-
     public ReadOnlyObservableCollection<ModCursor> Mods { get; }
-    public LoadoutId Loadout { get; set; }
+    public LoadoutId LoadoutId { get; set; }
+    public string LoadoutName { get; }
 
     public ReadOnlyObservableCollection<IDataGridColumnFactory> Columns { get; }
+
+    public Task AddMod(string path);
 
 }
