+++ src/NexusMods.App.UI/RightContent/LoadoutGrid/LoadoutGridView.axaml.cs
@@ -4,6 +4,7 @@ using Avalonia.Controls;
 using Avalonia.Data;
 using Avalonia.Markup.Xaml;
 using Avalonia.Markup.Xaml.Templates;
+using Avalonia.Platform.Storage;
 using Avalonia.ReactiveUI;
 using NexusMods.App.UI.RightContent.LoadoutGrid.Columns;
 using ReactiveUI;
@@ -17,13 +18,16 @@ public partial class LoadoutGridView : ReactiveUserControl<ILoadoutGridViewModel
         InitializeComponent();
         this.WhenActivated(d =>
         {
-            this.WhenAnyValue(view => view.ViewModel!.Toolbar)
-                .BindToUi(this, view => view.ToolbarViewHost.ViewModel)
-                .DisposeWith(d);
-
             this.WhenAnyValue(view => view.ViewModel!.Mods)
                 .BindToUi(this, view => view.ModsDataGrid.Items)
                 .DisposeWith(d);
+
+            this.WhenAnyValue(view => view.ViewModel!.LoadoutName)
+                .BindToUi(this, view => view.ModlistNameText.Text)
+                .DisposeWith(d);
+
+            AddModButton.Command =
+                ReactiveCommand.CreateFromTask(AddMod);
 
             this.WhenAnyValue(view => view.ViewModel!.Columns)
                 .OnUI()
@@ -39,6 +43,28 @@ public partial class LoadoutGridView : ReactiveUserControl<ILoadoutGridViewModel
                 })
                 .DisposeWith(d);
         });
+    }
+    
+    private async Task AddMod()
+    {
+        var provider = TopLevel.GetTopLevel(this)!.StorageProvider;
+        var options =
+            new FilePickerOpenOptions
+            {
+                Title = "Please select a mod file to install.",
+                AllowMultiple = true,
+                FileTypeFilter = new[]
+                {
+                    new FilePickerFileType("ZIP") {Patterns = new [] {"*.zip"}},
+                    new FilePickerFileType("RAR") {Patterns = new [] {"*.rar"}},
+                    new FilePickerFileType("7-Zip") {Patterns = new [] {"*.7z"}},
+                }
+            };
+
+        foreach (var file in await provider.OpenFilePickerAsync(options))
+        {
+            await ViewModel!.AddMod(file.Path.LocalPath);
+        }
     }
 }
