--- src/NexusMods.App.UI/Toolbars/IToolbarViewModel.cs
@@ -1,7 +1,0 @@
-﻿namespace NexusMods.App.UI.Toolbars;
-
-public interface IToolbarViewModel : IViewModelInterface
-{
-
-
-}
